Add tests for checkBrazeDependencies
Plus a bit of refactoring to support this. Moving some of the dependency
fetching (e.g. getBrazeUuid) out of StickyBottomBanner simplifies things
in there a bit.

I’m not much of a fan of all the mocking that we’re doing in the new
tests I’ve added but I think it’s worth doing for confidence around this
core bit of functionality, and at least it’s all centralised in this one
test, which I’m thinking of as more of an integration test.