Stop using dependencyChecker
As an experiment we wanted to see how checkBrazeDependencies looks
without using dependencyChecker. We have concerns about whether the
benefits of using dependencyChecker (mostly around consistency)
outweight the drawbacks (complex code to maintain and understand, which
realistcally won't be re-usable).

Note that I've expanded the tests around checkBrazeDependencies a little
to test the additional fields we expect to be returned, since we can no
longer rely on getting this "for free" with dependencyChecker.

Also note that the dependencyChecker approach kicked off all checks "in
parallel" which we're not doing here. Realistically this probably won't
make much difference because of JS's concurrency model, except perhaps
in the case of fetching the UUID since that involves IO. We could make
optimisations to ensure we kick off things like the UUID fetch earlier
but this will likely harm readability so decisions should be weighed up
accordingly.