Refactor this to use a loop over the dependencies
I chatted with William and discussed some alternatives, including an
approach which used Promise.allSettled. However I wasn’t happy with the
version using allSettled (it also isn’t available in the version of node
we’re on which introduces further friction). I’ve left the allSettled
version in here commented, but will delete in a subsequent commit.

I think what we’ve got now could be a happy medium between the original
dependencyChecker work and the if/else approach.